update Role Module

+++ ant-framework/src/main/java/com/framework/modules/sys/controller/MenuController.java
@@ -48,7 +48,7 @@ public class MenuController {
     }
 
     @PostMapping("/menus/delete")
-    public R deleteMenu(@RequestParam Integer id){
+    public R deleteMenu(@RequestParam Long id){
          menuService.deleteMenu(id);
         return R.ok();
     }
